# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teehive API Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/credits": {
            "get": {
                "tags": [
                    "Credits"
                ],
                "summary": "Get credits of current user",
                "operationId": "App\\Http\\Controllers\\Api\\CreditController::getCredits",
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/add-credits": {
            "put": {
                "tags": [
                    "Credits"
                ],
                "summary": "Add credits of current user",
                "operationId": "App\\Http\\Controllers\\Api\\CreditController::addCredits",
                "requestBody": {
                    "description": "Pass number of credits",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "credits"
                                ],
                                "properties": {
                                    "credits": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/deduct-credits": {
            "put": {
                "tags": [
                    "Credits"
                ],
                "summary": "Deduct credits of current user",
                "operationId": "App\\Http\\Controllers\\Api\\CreditController::deductCredits",
                "requestBody": {
                    "description": "Pass number of credits",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "credits"
                                ],
                                "properties": {
                                    "credits": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}